# nodejs-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "A BCF API"
  version: "1.0.0"
  title: "BCFAPI"
host: "virtserver.swaggerhub.com"
basePath: "/swisx/bcfAPI/1.0.0"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /bcf/{version}/projects/{project_id}/topics/{topic_guid}/comments:
    get:
      tags:
      - "comment"
      summary: "Retrieve a collection of all comments related to a topic"
      description: "Retrieve a collection of all comments related to a topic (default\
        \ ordering is date)."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidCommentsGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/comment_GET"
          examples:
            application/json:
            - guid: "C4215F4D-AC45-A43A-D615-AA456BEF832B"
              date: "2016-08-01T12:34:22.409Z"
              author: "max.muster@example.com"
              comment: "Clash found"
              topic_guid: "B345F4F2-3A04-B43B-A713-5E456BEF8228"
              authorization:
                comment_actions:
                - "update"
            - guid: "A333FCA8-1A31-CAAC-A321-BB33ABC8414"
              date: "2016-08-01T14:24:11.316Z"
              author: "bob.heater@example.com"
              comment: "will rework the heating model"
              topic_guid: "B345F4F2-3A04-B43B-A713-5E456BEF8228"
      x-swagger-router-controller: "Comment"
    post:
      tags:
      - "comment"
      summary: "Add a new comment to a topic."
      description: "Add a new comment to a topic. This operation is only possible\
        \ when the server returns the createComment flag in the Topic authorization."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidCommentsPOST"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      - in: "body"
        name: "comment_POST"
        required: false
        schema:
          $ref: "#/definitions/comment_POST"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/comment_GET"
          examples:
            appllication/json:
              guid: "A333FCA8-1A31-CAAC-A321-BB33ABC8414"
              date: "2016-08-01T14:24:11.316Z"
              author: "bob.heater@example.com"
              comment: "will rework the heating model"
              topic_guid: "B345F4F2-3A04-B43B-A713-5E456BEF8228"
      x-swagger-router-controller: "Comment"
  /bcf/{version}/projects/{project_id}/topics/{topic_guid}/comments/{comment_guid}:
    get:
      tags:
      - "comment"
      summary: "Retrieve a collection of all comments related to a topic"
      description: "Retrieve a collection of all comments related to a topic (default\
        \ ordering is date)."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidCommentsComment_guidGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      - name: "comment_guid"
        in: "path"
        description: "Comment guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/comment_GET"
          examples:
            application/json:
              guid: "A333FCA8-1A31-CAAC-A321-BB33ABC8414"
              date: "2016-08-01T14:24:11.316Z"
              author: "bob.heater@example.com"
              comment: "will rework the heating model"
              topic_guid: "B345F4F2-3A04-B43B-A713-5E456BEF8228"
      x-swagger-router-controller: "Comment"
    put:
      tags:
      - "comment"
      summary: "Update a single comment"
      description: "Update a single comment, description similar to POST. This operation\
        \ is only possible when the server returns the update flag in the Comment\
        \ authorization."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidCommentsComment_guidPUT"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      - name: "comment_guid"
        in: "path"
        description: "Comment guid"
        required: true
        type: "string"
      - in: "body"
        name: "comment_PUT"
        required: false
        schema:
          $ref: "#/definitions/comment_PUT"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/comment_GET"
          examples:
            application/json:
              guid: "A333FCA8-1A31-CAAC-A321-BB33ABC8414"
              date: "2016-08-01T14:24:11.316Z"
              author: "bob.heater@example.com"
              modified_date: "2016-08-01T19:24:11.316Z"
              modified_author: "bob.heater@example.com"
              comment: "will rework the heating model and fix the ventilation"
              topic_guid: "B345F4F2-3A04-B43B-A713-5E456BEF8228"
      x-swagger-router-controller: "Comment"
    delete:
      tags:
      - "comment"
      summary: "Deletes a single comment."
      description: "Deletes a single comment. This operation is only possible when\
        \ the server returns the delete flag in the Comment authorization."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidCommentsComment_guidDELETE"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      - name: "comment_guid"
        in: "path"
        description: "Comment guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Comment"
...
  /bcf/{version}/projects/{project_id}/topics/{topic_guid}/document_references:
    get:
      tags:
      - "document reference"
      summary: "Retrieve a collection of all document references to a topic."
      description: "Retrieve a collection of all document references to a topic."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidDocument_referencesGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/document_reference_GET"
          examples:
            application/json:
            - guid: "212ab37a-6122-448e-86fc-86503183b520"
              url: "http://example.com/files/LegalRequirements.pdf"
              description: "The legal requirements for buildings."
            - guid: "6cbfe31d-95c3-4f4d-92a6-420c23698721"
              document_guid: "472ab37a-6122-448e-86fc-86503183b520"
              description: "The building owners global design parameters for buildings."
      x-swagger-router-controller: "DocumentReference"
    post:
      tags:
      - "document reference"
      summary: "Retrieve a collection of all document references to a topic."
      description: "Retrieve a collection of all document references to a topic."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidDocument_referencesPOST"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      - in: "body"
        name: "Document_reference_POST"
        required: false
        schema:
          $ref: "#/definitions/Document_reference_POST"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/document_reference_GET"
          examples:
            application/json:
            - guid: "135ab37a-6122-448e-86fc-86503183b520"
              url: "http://example.com/files/LegalRequirements_Update.pdf"
              description: "The legal requirements for buildings."
      x-swagger-router-controller: "DocumentReference"
...
  /bcf/{version}/projects/{project_id}/topics/comments/events:
    get:
      tags:
      - "comment events"
      summary: "RRetrieve a collection of comment events related to a project."
      description: "Retrieve a collection of comment events related to a project (default\
        \ sort order is date)."
      operationId: "bcfVersionProjectsProject_idTopicsCommentsEventsGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/comment_event_GET"
          examples:
            application/json:
            - comment_guid: "C4215F4D-AC45-A43A-D615-AA456BEF832B"
              topic_guid: "A211FCC2-3A3B-EAA4-C321-DE22ABC8414"
              date: "2014-11-19T14:24:11.316Z"
              author: "Architect@example.com"
              events:
              - type: "comment_created"
                value: null
            - comment_guid: "C4215F4D-AC45-A43A-D615-AA456BEF832B"
              topic_guid: "A245F4F2-2C01-B43B-B612-5E456BEF8116"
              date: "2013-10-21T17:34:22.409Z"
              author: "Architect@example.com"
              events:
              - type: "viewpoint_updated"
                value: "b24a82e9-f67b-43b8-bda0-4946abf39624"
      x-swagger-router-controller: "CommentEvents"
  /bcf/{version}/projects/{project_id}/topics/{topic_guid}/comments/{comment_guid}/events:
    get:
      tags:
      - "comment events"
      summary: "RRetrieve a collection of comment events related to a project."
      description: "Retrieve a collection of comment events related to a project (default\
        \ sort order is date)."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidCommentsComment_guidEventsGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      - name: "comment_guid"
        in: "path"
        description: "Comment guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/comment_event_GET"
          examples:
            application/json:
            - comment_guid: "C4215F4D-AC45-A43A-D615-AA456BEF832B"
              topic_guid: "A211FCC2-3A3B-EAA4-C321-DE22ABC8414"
              date: "2014-11-19T14:24:11.316Z"
              author: "Architect@example.com"
              events:
              - type: "comment_created"
                value: null
            - comment_guid: "C4215F4D-AC45-A43A-D615-AA456BEF832B"
              topic_guid: "A245F4F2-2C01-B43B-B612-5E456BEF8116"
              date: "2013-10-21T17:34:22.409Z"
              author: "Architect@example.com"
              events:
              - type: "comment_text_updated"
                value: "This is the updated comment"
      x-swagger-router-controller: "CommentEvents"
definitions:
...
  document_reference_GET:
    type: "object"
    required:
    - "guid"
    properties:
      guid:
        type: "string"
      document_guid:
        type: "string"
      url:
        type: "string"
      description:
        type: "string"
    description: "Schema for a single document reference GET, BCF REST API."
...
  comment_event_GET:
    type: "object"
    required:
    - "actions"
    - "author"
    - "comment_guid"
    - "date"
    - "topic_guid"
    properties:
      comment_guid:
        type: "string"
      topic_guid:
        type: "string"
      date:
        type: "string"
      author:
        type: "string"
      actions:
        type: "array"
        items:
          $ref: "#/definitions/event_action"
        minItems: 1
...
  comment_GET:
    type: "object"
    required:
    - "author"
    - "comment"
    - "date"
    - "guid"
    - "topic_guid"
    properties:
      guid:
        type: "string"
      date:
        type: "string"
      author:
        type: "string"
      comment:
        type: "string"
      topic_guid:
        type: "string"
      viewpoint_guid:
        type: "string"
      reply_to_comment_guid:
        type: "string"
      modified_date:
        type: "string"
      modified_author:
        type: "string"
      authorization:
        $ref: "#/definitions/comment_GET_authorization"
...
  comment_actions:
    type: "array"
    items:
      type: "string"
      example: ""
  event_action:
    type: "object"
    required:
    - "type"
    properties:
      type:
        type: "string"
      value:
        type: "string"
...
  comment_POST:
    type: "object"
    required:
    - "comment"
    properties:
      comment:
        type: "string"
      viewpoint_guid:
        type: "string"
      reply_to_comment_guid:
        type: "string"
    example:
      comment: "will rework the heating model"
  comment_PUT:
    type: "object"
    required:
    - "comment"
    properties:
      comment:
        type: "string"
      viewpoint_guid:
        type: "string"
      reply_to_comment_guid:
        type: "string"
    example:
      comment: "will rework the heating model and fix the ventilation"
...
  Document_reference_POST:
    type: "object"
    properties:
      document_guid:
        type: "string"
      url:
        type: "string"
      description:
        type: "string"
    example:
    - url: "http://example.com/files/LegalRequirements.pdf"
      description: "The legal requirements for buildings."
...
  comment_GET_authorization:
    properties:
      comment_actions:
        $ref: "#/definitions/comment_actions"
...